TSD destruction magic -- take 2.
When the TSD destructor is called and the AsanThread object corresponding to the current 
thread is destroyed, we write an odd integer value into the TSD slot.
This is invisible to other classes, because they call AsanThreadRegistry::GetCurrent() (which still returns
NULL for this case)
instead of directly accessing the TSD via pthread_getspecific().
In order to check that the current thread's AsanThread object has been destroyed,
another class may call AsanThreadRegistry::IsCurrentThreadDying()
For example, we don't fail now in pthread_create() if GetCurrent()==NULL and IsCurrentThreadDying()==true
(in this case we're setting the parent TID to -1)