dr_asan: Use dr_get_proc_address instead of drsyms, and other comments and simplifications.
For now, only instrument base+index+displacement memory operands, and ignore
TLS and absolute references.  This where 99.9% of the bugs will be.

Simplify the process for picking R2.  We only need to make sure it doesn't
conflict with R1 or the registers used by the operand.

Added the Windows x64 regparm, but it needs more than that (alignment and
built-in spill space).