Move the module checks out of the bb event and instrument more code.
TBR by Timur after holiday.

I moved the module checks to the module load event and then stored the
ranges in a sorted vector.  The goal was to speed up the common case of
translating bbs from the app that are already asan-instrumented, but
this didn't have any effect.  I removed the module checks that I didn't
need to run net_unittests: libpthread, libdynamrio, libdrpreload, etc.

We still can't instrument libc and JITed code, which are big blindspots
IMO.