Tell DR not to instrument our ignored modules.
This also skips instrumenting buffer overreads in *libX*, since we found
one in XChangeProperty.

We also avoid re-instrumenting libppGoogleNaClPluginChrome.so, which is
compiler instrumented.  Long term, we can probably check if a module
imports __asan_init to see if it's been instrumented or not.

Also had to fix a bug in call reachability on x64 with -fpie.