kasan_symbolize: don't print inlined-to function for inlined frames
Currently inlined function name is in the middle of line with lots of text around,
it makes it very difficult to find it visually.

Don't print inlined-to function name for inlined frames,
it makes it much easier to find inlined function name visually.
This info is redundant anyway. Compare:

Before:
     [<     inlined    >] __mutex_init+0x69/0x80 atomic_set ./arch/x86/include/asm/atomic.h:47
     [<     inlined    >] __mutex_init+0x69/0x80 osq_lock_init include/linux/osq_lock.h:29
     [<ffffffff810de949>] __mutex_init+0x69/0x80 kernel/locking/mutex.c:59

After:
     [<     inlined    >] atomic_set ./arch/x86/include/asm/atomic.h:47
     [<     inlined    >] osq_lock_init include/linux/osq_lock.h:29
     [<ffffffff810de949>] __mutex_init+0x69/0x80 kernel/locking/mutex.c:59